efs/requirements.txt
certifi==2020.6.20
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.7
django-crispy-forms
django-mathfilters==1.0.0
djangorestframework==3.11.0
gunicorn==20.0.4
psycopg2==2.8.4
django-heroku
django-bootstrap4
twilio==6.32.0
pisa
xhtml2pdf